bug(files and folder): rename file and folder [Updates #165657411]
