+++ server/app/middleware/statusCheck.js
