+++ server/app/migration/createTable.js
