+++ server/index.js
@@ -2,9 +2,9 @@ import express from 'express';
 import bodyParser from 'body-parser';
 
 
-import userWithDb from './usingDB/route/userRouteWithDB';
-import accountWithDB from './usingDB/route/acctRouteWithDB';
-import transactionWithDB from './usingDB/route/transactRouteWithDB';
+import userRoute from './app/route/userRoute';
+import accountRoute from './app/route/acctRoute';
+import transactionRoute from './app/route/transactRoute';
 
 
 const app = express();
@@ -18,9 +18,9 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.get('/', (req, res) => res.send('Welcome to Banka'));
 
 
-app.use('/api/v1/auth', userWithDb);
-app.use('/api/v1', accountWithDB);
-app.use('/api/v1/transactions', transactionWithDB);
+app.use('/api/v1/auth', userRoute);
+app.use('/api/v1', accountRoute);
+app.use('/api/v1/transactions', transactionRoute);
 
 
 app.use('*', (req, res) => {
