--- server/usingDB/route/userRouteWithDB.js
@@ -1,12 +1,0 @@
-import express from 'express';
-import dbUserController from '../controllers/user-controller';
-import { signUpValidation, signInValidation, createStafValidation } from '../middleware/userValidation';
-import authorize from '../middleware/userAuthorization';
-
-const dbUserRoute = express.Router();
-
-dbUserRoute.post('/signup', signUpValidation, dbUserController.signupUser);
-dbUserRoute.post('/signin', signInValidation, dbUserController.signinUser);
-dbUserRoute.post('/create', createStafValidation, authorize.verifyAdmin, dbUserController.createStaff);
-
-export default dbUserRoute;
